pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omnidata-nexus-core"
version = "1.1.0"
description = "Modular, asynchronous DataLoader framework for financial and macroeconomic data aggregation"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "OmniData Nexus Core Team"}
]
keywords = ["finance", "data", "async", "api", "fmp", "polygon", "fred", "dataloader"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
    "Topic :: Office/Business :: Financial :: Investment",
    "Framework :: AsyncIO",
    "Typing :: Typed",
]

dependencies = [
    "aiohttp>=3.9.0,<4.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "typing-extensions>=4.0.0,<5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<9.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-cov>=4.1.0,<6.0.0",
    "aioresponses>=0.7.4,<1.0.0",
    "mypy>=1.5.0,<2.0.0",
    "ruff>=0.1.0,<1.0.0",
    "detect-secrets>=1.4.0,<2.0.0",
]
docs = [
    "mkdocs>=1.5.0,<2.0.0",
    "mkdocs-material>=9.4.0,<10.0.0",
    "mkdocstrings[python]>=0.24.0,<1.0.0",
]

[project.urls]
Homepage = "https://github.com/Maeshowe/Nexus_Core"
Documentation = "https://maeshowe.github.io/Nexus_Core/"
Repository = "https://github.com/Maeshowe/Nexus_Core.git"
Issues = "https://github.com/Maeshowe/Nexus_Core/issues"
Changelog = "https://github.com/Maeshowe/Nexus_Core/blob/main/docs/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (may use real APIs)",
    "e2e: End-to-end tests (full workflow)",
    "slow: Tests that take longer to run",
    "security: Security tests (API key sanitization, etc.)",
    "fmp: Tests for FMP provider",
    "polygon: Tests for Polygon provider",
    "fred: Tests for FRED provider",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
strict_optional = true

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "SIM102", # nested if - often more readable
    "SIM103", # return condition directly - style preference
    "SIM108", # ternary operator - style preference
    "ARG001", # unused function argument - fixtures have side effects
    "ARG002", # unused method argument - fixtures have side effects
]

[tool.ruff.lint.isort]
known-first-party = ["data_loader"]
